requirements.txt
tensorflow>=2.10.0
numpy
Pillow
opencv-python
pytesseract
regex
tk